move user queries to separeted directory
